s/vector?/sequential?/ when applying transform functions
The transform function code decides whether or not to use `map` by
testing with `vector?`, which breaks when the result set is a seq
instead, as is the case when you do a query of the form
`(select foos (with bars))`. As far as I can tell this could be
fixed either by replacing the `vector?` test with a `sequential?`
test, or using a `mapv` in `apply-posts` (and maybe elsewhere).

This commit implements the former, and adds a test.